EnemyPrefab now reads form EnemyStats, Spawner now initializes enemies, enemies now despawn when they reach the target and end wave gets called
Set up everything that would need to occur for the first wave to spawn and happen for enemies, NOTE: Enemies now need to be initialzied with MobStats in order to function. If an enemy does not have mobstats it cannot function. Assign MobStats either automatically from the spawner or manually by InitalizeOnAwake = true and placing in a mobstats